fixed the quotation handling
Putting quote around message is wrong on Unix, where quote is handled by shell. But on Windows, where all t he arguments need to be packed to a single String before the kernel API call, quotation can behave differently from one runtime to another.

Therefore it is better not to trust that too much. Instead, opt for passing it in as a file.