[FIXED JENKINS-20684] Parser for .netrc file format
This path implements a parser for the .netrc file supporting various formats.
It is not restricted to having a machine definition on a single line but can
parse any valid format. The one exception is that it does *not* cope with
`macdef` definitions.

handle `macdef` definitons in the netrc file to the parser.
Anything following a `macdef` keyword will be ignored up to
the next empty line.