Create new class Netrc for .netrc file handling.
Create new class Netrc which does all the .netrc file handling, i.e.
it parses the file and stores all machine accounts in a map from which
it can then provide credentials for a specific host.
This can be enhanced in the future to return reused instances for the
same file, which it currently doesn't.