[JENKINS-19994] Close files opened by JGit operations
The JGit Repository object opens files and requires that the close()
method on the repository object is called to close the files.

The JGit RevWalk object documentation states that the dispose()
method should be called on the RevWalk object to unlock any resources
it is holding.

The JGit ObjectReader documentation status that its resources should
be released by calling the release() method.

This change is not enough to allow workspace cleanup in all cases, since
there are other cases (outside the JGit implementation) which will cause
files to be kept open in the workspace.  Refer to JENKINS-20585 for one
example.