Fix JGit getDefaultRemote (deprecated) implementation to match CliGit
Added several tests in the deprecated test methods to assure the
implementations are consistent with one another.

There is a variation between CliGitAPIImpl and JGitAPIImpl ordering
which is handled in the case where more than one remote is defined.
The javadoc for the CliGitAPIImpl says that it returns the first
remote defined, though it seems to actually return the last remote
defined.  The JGitAPIImpl returns the first remote defined so that
it is consistent with the javadoc.