Compare checkout of ambiguous reference between ClitGit and JGit
Test shows that when a tag and a branch have the same name, CliGit
(using git 1.7.10.4) prefers to checkout the branch, while JGit (using
JGit 3.2.0) prefers to checkout the tag.

Since the reference is ambiguous, it seems quite reasonable that
different implementations might have different results when referring
to the ambiguous reference.