Construct a Repository object instead of using FileBasedConfig for JGit config
All my attempts to read useful results through the FileBasedConfig
implementation failed.  Using a RepositoryBuilder with the GIT_DIR
argument to set the working tree allows one or more of the test cases
to pass.

There are still differences between the JGit and CliGit
implementations, but this reduces those differences.