Set remote URL in CliGit clone method, match JGit and command line
The CliGit clone_ method is implemented internally as fetch plus merge,
so that credentials can be used with command line git.  Unfortunately,
that means the results of CliGit clone_ are not the same as command line
git clone, and not the same as JGit clone.

This change sets the remote URL so that CliGit and JGit are the same in
that area.

This does not perform a checkout of the master branch after the clone,
since the typical consumers of a clone will perform their own checkout
of a specific SHA1.