Test prune with pathologically named branches
If a branch is created and pushed to a remote repository (for example,
a branch named "branch1"), then that branch is deleted, and a new
branch is created and pushed which uses the first branch name as a
parent (for example, a branch named "branch1/a"), the fetch will fail
until a "git remote prune origin" has been performed to remove the
originally created remote branch name from the working repository.

Refer to https://issues.jenkins-ci.org/browse/JENKINS-19591 for a more
detailed description.

This is another case where the JGit implementation differs from the
command line git implementation.  The JGit implementation does not
seem to require the "git remote prune origin" command.  I assume it
performs that operation happens implicitly inside JGit.