Test deprecated push(RemoteConfig, String) method
The test shows that there is a problem in the JGit revParse
implementation when used with multiple WorkingArea objects within the
same test.  The same problem does not afflict the CliGit
implementation.  The problem is noted with an "instanceof"
conditional, just as a few of the other differences between JGit and
CliGit are noted.