tl;dr: s/UsernamePasswordCredentialsImpl/StandardUsernamePasswordCredentials/
NOTE: This is the same change that I had made previously in this pull request: https://github.com/jenkinsci/git-client-plugin/pull/32

I am not sure how or why my "merged" pull request does not appear in the commit history of the repo.

The original notes from PR#32 follow:

Weaken the type requirement on the instanceof used in the git-client plugin to only rely on the object being a StandardUsernamePasswordCredentials.

The interface is both necessary and sufficient for the APIs used.

This is morally equivalent to how the private key logic is already using the SSHUserPrivateKey interface vs. a specific implementation.