Add isBareRepository tests
The assertions are intended to detect unintentional changes to a
deprecated API.  It is safe to assume that since the API is
deprecated, its behavior should not change with future plugin
releases.

The JGit implementation attempts to match the CliGit implementation
but doesn't quite match.  The prioritization of assertions in this file
attempts to show the most critical use cases for isBareRepository are
covered in JGit and CliGit with similar results.

The Javadoc for the CliGit implementation at one time stated that
its argument must be the ".git" directory.  If the caller passes
the .git directory (for a working directory) or the root directory
of the repository (for a bare repository), then the JGit and CliGit
implementations return consistent results.

Fix WorkingArea repo() method for bare repo