Fix deprecated JGit.setRemoteUrl(String, String, String)
Was not being tested from the unit tests, and when called was not
performing the expected operation.  The implementation now uses the
same pattern as JGit.setRemoteUrl(String, String).  That is now called
by unit tests, and passes the tests.

Also added calls to submoduleSync() without any assertions of the
results of the call.  Increases coverage slightly and will detect if
submoduleSync() throws an unexpected exception.  Further detection of
submodule errors will need to wait until the submodule tests have been
further extended.