Add test of msysgit 1.9 core.longpaths setting
The msysgit 1.9 release includes a new core setting, core.longpaths,
which allow content file names longer than 256 characters.  The option
is only available with msys git.  There is currently no support in
git-client-plugin for the longpaths setting.  This is only a test to
confirm that the longpaths setting in msysgit 1.9 works as expected,
in case the plugin chooses to use that setting in the future.

The JGit implementation already works with long paths.

The Linux C git implementation already works with long paths.