Enable getAllLogEntries test again - fixed
The getAllLogEntries method seems to return a time ordered list of
commits, but the granularity of the time ordering seems to be only to
a single second.  The namespace test creates three branches very
rapidly, each with a distinct commit.  That seems to highlight a
difference in the time ordering of the results of C git and JGit
getAllLogEntries.

Work around is to clone from the original repository with a reference
to the local mirror, rather than cloning from the local mirror
directly.