Make isAtLeastVersion package visible
Make isAtLeastVersion package visible so that it can be tested properly.  To
test this method properly, we need a new method (computeGitVersion) to force
the internal Git version to test various different versions.  Inspiration by
MarkEWaite/git-client-plugin/13879a5.

Also, since isAtLeastVersion is now package visible, we can get rid of the
copy of version code in GitAPITestCase and use this method.