Switch to components of powermock rather than full package
Using the powermock full release package introduces a dependency on
TestNG which causes the maven output to include a warning that no
TestNG classes were found.  This project uses JUnit, so no TestNG
classes are ever expected to be found.

http://stackoverflow.com/questions/1796833 suggests that the simplest
solution is to use components of PowerMock explicitly rather than
using the full release package.