Update severl maven properties so that 'mvn site' succeeds
The parent pom provides outdated versions of the dependency plugin,
the enforcer plugin, the project-info-reports plugin, and the site
plugin.  This overrides those outdated versions with the most current
versions, along with a big comment that will remind us to remove the
override if the parent ever updates to newer versions.