[Fixed JENKINS-21806] Improve ssh.exe finding on Windows with msysgit
Adapt to msysgit installed in non-default locations, and to msysgit
alternative configurations as allowed by the msysgit installer.  The
mssygit installer allows the user to choose to make no changes to the
PATH, or to add the "cmd" directory to the PATH, or add the "bin"
directory to the PATH.

If the user chooses to make no changes to the PATH, then they
configure their Jenkins git with the absolute path of the git
executable.

If the user chooses to add the "cmd" directory to the PATH, then they
configure their Jenkins git with the executable as "git", "git.exe",
or (with some msysgit versions) "git.cmd".

If the user chooses to add the "bin" directory to the PATH, then they
configure similarly to the "cmd" directory case.

The plugin attempts to handle all those cases correctly.

Special thanks to Thiago Zanetti for providing the original code for
this fix.