ls-remote to lookup references in a remote repository
List references, like heads or tags, in a remote repository
without depending on a locally cloned workspace.

The git-parameter plugin lists branches or tags on a local
cloned workspace, so it doesn't work on the first build or
if user wipes out workspace.

If user only wants a tag name with its commit id, there is
no need to clone the whole repository.