Update JGit dependency to 3.5.3 - security fix in JGit
Refer to http://article.gmane.org/gmane.linux.kernel/1853266 for a
description of the client side bug which is resolved by JGit 3.5.3.

It seems unlikely the client side bug will affect many Jenkins users
due to the relatively limited workflow used in the git client plugin
and the git plugin.