Send stderr, stdout, and exception messages to CliGitAPIImpl listener
Reduce the number of cases which require guesses about root causes of
command line git failures.

Previously, when command line git failed (with an IOException, for
example), the exception message was not presented to the user, even
though it includes crucial hints about the root cause of the exception.
This change provides the listener with the exception message (not the
exception stack trace, just the message).