Revert "Send stderr, stdout, and exception messages to CliGitAPIImpl listener"
Always reporting the stderr and stdout of failed command line git calls
inserts extra noise into the console output when those command line calls
are expected to fail.  For example, "git config core.sparsecheckout" is
used to read the value of core.sparsecheckout. If no value is set, then
the command line call returns a non-zero value which generates stderr
and stdout into the console log, even though the error is expected and
is correctly handled in the caller.

Need some better way of increasing and decreasing the diagnostic output
of the git-client, rather than always showing failures.  Need to use the
existing logging facility which can be increased and decreased as needed,
and managed from the Jenkins logging console.

This reverts commit 744cb9fef13fd57d7f616c5ed27b0092aefc8ba3.