Fix failure in test_push_tags
Currently (with 1.16.1), when testing git-client-plugin, I am getting
the following error in test_push_tags:

--------------------------------------------------------
The current branch master has no upstream branch.
To push the current branch and set the remote as upstream, use

    git push --set-upstream /tmp/hudson4099265862244982437test master
--------------------------------------------------------

The reason that this is happening is that test_push_tags() is using the
PushCommand() pattern, but is failing to specify a refspec.  Also, this
test case is creating a remote url, but is not setting up the master
branch to track that remote url.  So, without a refspec, the behaviour
depends on whether CLI or JGIT is used:
   -- CLI -- it uses whatever push.default specifies.
      In my case, it was set to tracking/upstream.  With this setting,
      it requires that the upstream be configured correctly, or it will
      error out with the above error message.

   -- JGIT -- it uses Transport.REFSPEC_PUSH_ALL
      This corresponds to push.default = matching

If you are wondering why this was not hit before, it depends on
push.default defaulting to matching.  Any other value casues the above
failure in the CLI GIT implementation.  Also, all other instances of
git.push() in the test harness uses the 2 argument version, which
requires the refspec().  Hence, no problem.

To make this test deterministic, specifying the refspec via
.ref("master"). fixes it for both configurations.