Base 1.17.1 off 1.16.1: allow more study of git 2.3.6-SNAPSHOT issues
The release of git client plugin 1.17.0 delivered a fix for
[ISO date parsing](https://issues.jenkins-ci.org/browse/JENKINS-27097).
Unfortunately, that fix exposes a long-standing bug in git plugin versions
2.3.5 and earlier.  The bug is fixed in git plugin 2.3.6-SNAPSHOT.
However, tests of git plugin 2.3.6-SNAPSHOT show that it is not yet
ready for production use.

Rather than rush to deliver git plugin 2.3.6, this change will restore the
git client plugin to the code base which was delivered with git client
plugin 1.16.1. That will hide the ISO date parsing bug that is in git
plugin 2.3.5 and earlier and allow more time to prepare 2.3.6 for release.

The 1.17.0 changes are *not* included in this build, including:

* Update minimum Jenkins version from 1.509 to 1.532.2
* Update JGit to 3.7.0
* [Allow prune with JGit JENKINS-26197](https://issues.jenkins-ci.org/browse/JENKINS-26197)
* [Use ISO-8601 date format in changelog JENKINS-27097](https://issues.jenkins-ci.org/browse/JENKINS-27097)
* [Improve checkout failure message JENKINS-26748](https://issues.jenkins-ci.org/browse/JENKINS-26748)
* Guard against null pointer exception in several API calls
* Update other dependencies