[Fix JENKINS-28943] Find windows ssh even if git is absolute path to cmd dir
The path guesser in getPathToExe correctly checks for cmd and exe
suffixes, and correctly checks each entry in the PATH environment
variable.  Prior to this change, it did not check for the absolute
path having been specified with the cmd directory rather than the bin
directory.

The user had set C:\Program Files\Git\cmd\git.exe as the git program
for that slave.  That is an allowed setting, and it executed most git
commands correctly. It failed on ssh based commands.