git-client: store username as default user in credentials.username
In case the http(s) url being used does not supply a username, set the
default username for credentials in the gitconfig. This helps ensure
that credentials which have usernames not matching the system username
of the slave will work as expected for http(s) urls.

JENKINS-29604 - git-client CLI implementation does not propogate username

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>