git-client: use username in the GIT_SSH setup
Teach the GIT_SSH script we create to pass "-l <username>" so that a
default user will be selected incase the URL does not encode the
username. This is important, as otherwise the CLI implementation will
default to the local username, which is probably incorrect. While users
are able to work around this bug by specifying the correct user in the
ssh url, they cannot do so for submodules.

JENKINS-29604 - git-client CLI implementation does not propogate username

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>