fix git describe test case for uneven commit abbreviation length
git describe will honor the "core.abbrev" setting from the gitconfig,
(thus CLI implementation will as well). This potentially breaks the JGit
test of git describe as implementations may not return the exact same
length for an abbreviated hash. Fix this by implementing
StringSharesPrefix matcher based on SubstringMatcher from hamcrest.

We can't use the startsWith matcher due to not knowing which particular
string will actually contain the shorter prefix. However, we want to
allow both cases as long as the total string continues to match up to
the end of either string.

This fixes a test failure if a user has customized their Git
configuration to provide a larger default abbreviation (such as for
working with incredibly large projects like the Linux kernel). We
shouldn't fail the test just because the JGit implementation returns
fewer characters.

Note that we will have to update StringSharesPrefix if we ever update
the hamcrest version requirement, as the upstream has changed.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>