remove invalid pushNoRefSpec test
This behavior is no longer allowed since something like Git version
1.7.10, as a new test was added to prevent push to a separate upstream,
ie:

git push <remote url>

without a refspec and without properly being setup as the remote for the
branch, push can't properly figure out what to push. It was bugged and
ended up pushing branch.<upstream>.merge to <remote> which is completely
bogus. See 135dadef712f ("push: error out when the "upstream" semantics
does not make sense") from the upstream git repository for more
information.

We could also theoretically change the test to actually pass the remote
upstream for this branch, but then we fall under the user configured
"push.default" which changed behavior under 2.0, and we should not rely
on these sorts of behaviors in the plugin. It is best to always be
explicit so that tests do not (normally) depend on the environment under
which they are performed.

Thus, remove the test of this behavior as it causes issues when running
a git client greater than 1.7.x for the CLI implementation. It is
possible JGit will succeed, but I do not believe this should be
supported behavior, and instead we should enforce the same behavior as
the CLI upstream git implementation, which requires proper setup of
"upstream" semantics.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>