rework tricky pushNoRefSpec test to be explicit
Do not test "no refspec" push behavior, as this is user configurable and
infact the default changed in Git 2.0 Despite this change, the new
default still functions with the supplied test, but we should not rely
on this behavior. Instead, re-write this test to be "pushCurrent" and
ensure we explicitly mention "HEAD" as the ref to push (rather than our
configured refspec). In this way, we can ensure this test succeeds even
if the CLI implementation has a different configured default push mode,
and we are actually testing what we think.

I considered removing the test altogether, however this breaks other
tests which depend on commits being pushed into the remote, which does
not have a 2nd commit until we complete this initial test. Due to this,
it is simpler to just redefine this test to work regardless of how the
user configured the default settings.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>