Fix flawed test interdependency in PushSimpleTest
The checkoutOldBranchAndCommitFile method must have at least two commits
available on the branch in order to perform the checkout of the old
commit.  Create a second commit in the bare repository on each branch
that is created so those two commits will be available.

Removes pushCurrent test, test case is already covered in PushTest.