git-client: use GIT_ASKPASS for all username/password credentials
Currently, we support username/password authentication only for http(s)
URLs. In addition, this support requires a credential store which is
only useful for fetch, and not for clone. Our implementation of
CloneCommand uses a combination of git-init and git-fetch for this very
reason.

Instead, add support for any set of UsernamePassword credentials via the
GIT_ASKPASS method. Write a script which compares the $1 argument
(passed in by Git in english as a suitable prompt). This script will
output the provided Username and Password as requested by the provided
prompt. The prompt starts with the correct term but may contain more
characters so limit the check only to the first 8 characters.

We also provide (very) limited protection against unknown symbols. I am
not certain if "^" will be interpreted in the windows shell, so I only
escaped %. For Unix shell, escape single quotes by ending the string,
inserting an explicit single quote, then restarting the string. This
should work for all strings as the only special character in a single
quote string is itself the single quote.

This should fix HTTP authentication for submodules, as well as SSH
username-password combinations which do not use a private key. In
addition, we escape the passphrase for unlocking the private key in the
SSH key phase, so that behavior is maintained across these types of
authentication.

Due to these changes, support for the credentials store and helper
functions associated with this change have also been removed.

It should now be possible to revert the use of init+fetch in
CloneCommand, but git-revert can't do it cleanly so I leave this as an
exercise for the future.

Finally, this patch also removes "-c core.askpass=true" as this is the
same setting as "GIT_ASKPASS". Although git documentation and code says
that it honors GIT_ASKPASS above core.askpass, setting askpass to true
is almost certainly wrong. core.askpass is not a boolean, but the same
as GIT_ASKPASS. That is, it should be set to the location of a program
which outputs the password.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>