git-client-plugin: have CredentialTest allow specifying the file.
Today, CredentialTest always searches for "README.md" which may not
actually be a file in the repository. Allow specifying the file in the
repos.json. No file will optionally fallback to the current default of
"README.md", and thus should make repos.json continue to work. This
allows for adding other repos to the repos.json file.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>