git-client-plugin: support attempting submodule in CredentialTest
This patch extends the repos.json format to include a "submodules"
boolean. If the submodules are desired for this repository you can add
submodules: true to the json file.

It is suggested that repos which rely on this should change their file
setting to check for a file located inside of the submodule, as this
will enable checking that submodules actually did get checked out.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>