jgit: add support for Credentials to JGit's Submodule Update command
JGit allows passing a CredentialsProvider into the SubmoduleUpdate
command. We need to rename our SubmoduleUpdateCommand references so that
we avoid namespace conflicts, but otherwise this works quite well. Using
a local submodule test in repos.json I get it to clone a submodule and
verify the particular file is available.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>