git-client: add support for use of parent project credentials
CliGit implementation uses this to decide whether it should lookup the
parent credentials assuming default remote. It will attempt to discover
the credentials of the parent and pass those for the submodule instead
of requiring that the git plugin provide credentials per URL. A future
version of the Git plugin will add support for this advanced feature of
submodules.

The JGit implementation is a no-op, because as far as I can tell JGit
passes a credentials Provider which should already work. If it doesn't
work, then I am not sure how to fix it.

To help with testing, add the option "parentcreds" to the JSON
credentials file which enables this feature when set to true.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>