Fixed garbled commit messages on Windows
On windows changelog commit messages with unicode characters are
not saved correctly to changelog.xml when CliGitAPI
implementation is in use.

That happens because "git whatchanged" gives byte stream of data.
Commit messages in that stream are encoded in UTF-8.  It is
necessary to explicitly decode bytestream to strings using UTF-8
encoding, otherwise default system encoding will be used.

This should fix issues:
  https://issues.jenkins-ci.org/browse/JENKINS-6203
  https://issues.jenkins-ci.org/browse/JENKINS-14798
  https://issues.jenkins-ci.org/browse/JENKINS-23091