[JENKINS-30371] Distribute JGit java7 jar to allow JGit symlink support
JGit symlink support is optional with JGit 3.7.1.  It requires a copy of
the JGit java7 jar, a Java 7 VM, and a file system which supports symbolic
links.  Including the jar file in a Java 6 environment is safe because
the JGit code won't load the additional jar when running on Java 6.