Disable ssh password and passphrase prompting
Since we're already using StrictHostKeyChecking=no, we should be able
to use BatchMode, another OpenSSH configuration setting to reduce the
changes of git blocking to prompt for a password.

The ssh_config man page for OpenSSH says:

BatchMode

If set to “yes”, passphrase/password querying will be disabled. This
option is useful in scripts and other batch jobs where no user is present
to supply the password. The argument must be “yes” or “no”. The
default is “no”.