Warn JGit can't shallow clone rather than throwing exception
Previous code warned when clone was requested with a shallow clone,
but then threw an unsupported operation exception if fetch was called
with shallow, even if the shallow request to fetch was that the fetch
not be shallow.

Will now warn rather than throw an exception if a shallow fetch is
requested or if depth is set for a shallow fetch. That makes shallow
and depth behavior for fetch consistent with clone.

JGit server support for shallow clone has been implemented, but JGit
support for shallow clone to clients is not yet implemented.  The plugin
needs shallow clone to clients.