git-client: add support for --force option to submodule update
Also add a test case for --force option

Since the supported versions of git do not all support "-C" syntax,
extend our cmd(), and launchCommand() syntax with a launchCommandIn()
syntax, which we will make use of. This compiled with w.file() allows us
to launch a command within a sub directory.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>