Prepare for release of 2.0.0-beta1
This is now proposing a larger change to no longer compile on
Java 6.  It will require at least Jenkins 1.625 and Java 7.

The change has been a very nice improvement in the JGit implementation,
since the code can now use JGit 4 and can use the Java 7 "try with
resources" constructs to assure that files are correctly closed on
exception cases.