Create missing directory in cli git.init() method
Multi-branch pipeline plugin and other AbstractGitSCMSource callers were
initially using JGit as their implementation, and developed an unexpected
dependency on JGit behavior. JGit init() (in JGit 3.7 at least) creates
the directory if it does not exist. Rather than change the multi-branch
pipeline when the git client plugin was adapted to allow either git or
jgit, instead the git.init() method was changed to create the target
directory if it does not exist.

Low risk from that change of behavior, since a non-existent directory
caused the command line git init() method to consistently throw an
exception.