Don't prune symbolic refs during JGit fetch
JGit prune implementation inside JGitAPIImpl deletes all remote branches
prior to the fetch, then relies on fetch() to recreate those branches.
As of JGit 4.5, the JGit FetchCommand no longer recreates the HEAD
reference during fetch.

Since HEAD is a symbolic reference, that is a reasonable optimization
inside JGit, and should be honored by JGitAPIImpl.

If users push symbolic refs to their upstream repository (as described
in a talk at Jenkins World 2016), they might not be pruned by this
implementation.

This could use an additional several test cases around symbolic references
coming from remote repositories. If a remote symoblic refernece is cloned
as a branch, then it will be pruned.  If a remote symbolic reference is
cloned as a symbolic reference, then that symbolic reference will never
be pruned (due to this change).