Silence unchecked cast compiler warning & method name error
The getApplicableDesccriptors method implementation causes the Java 8
compiler to report an unchecked cast warning, even though the cast is
guarded by an isAssignableFrom check.

The getApplicableDesccriptors method name has been corrected
to be getApplicableDescriptors, with a deprecated method named
getApplicableDesccriptors to assure compatibility in case there are
callers to the now incorrectly spelled method.