Use @NotImplementedInCliGit to exclude CliGit submodule tests
Don't test JGitAPIImpl for renamed submodules

Renamed submodule support is not implemented in JGit.
Rather than disable the entire test, this disables a single check in
the test.  The JGit implementation is still wrong, and this only
checks that it is no more wrong than it was before the
tests/getSubmodules branch received a renamed submodule.