More submodule test improvements
Several of the submodule tests were not asserting enough conditions. This
commit adds additional assertions, and includes a few JGit tests in the
submodule test cases.

JGit does not seem to handle renamed submodules, so some of the tests
are not executed with JGit.  There are other cases where JGit behaves
differently than command line git.  Tests are adapted to show those
differences.