Adapt Cli.setDefaults to git config command failure
The "git config" command returns a failure if the config key does not
exist.  The CliGitCommand previously asserted that all command line git
calls must succeed.  That is a good general pattern for a test author,
since a failing call to command line git will most often mean that the
test author made a mistake in the command line git call.

Added a private method inside CliGitCommand which can accept a failed
call and still return results from the call.